backend/requirements.txt
Django>=5.1
djangorestframework>=3.15
django-cors-headers>=4.3
django-environ>=0.11
mysqlclient>=2.2
Pillow>=11.0
django-filter>=24.3
celery>=5.4
redis>=5.2
python-telegram-bot>=21.9
openai>=1.59
anthropic>=0.42
requests>=2.32
gunicorn>=23.0
django-celery-beat>=2.7
